+++ soc_perf/include/server/socperf_server.h
@@ -35,7 +35,7 @@ public:
     void PowerLimitBoost(bool onOffTag, const std::string& msg) override;
     void ThermalLimitBoost(bool onOffTag, const std::string& msg) override;
     void LimitRequest(int32_t clientId,
-        const std::vector<int32_t>& tags, const std::vector<int32_t>& configs, const std::string& msg) override;
+        const std::vector<int32_t>& tags, const std::vector<int64_t>& configs, const std::string& msg) override;
 
 public:
     SocPerfServer(int32_t systemAbilityId, bool runOnCreate);
